added: TransitionTypes | refactor: Work(U32F32) instead of f64 for work fields

+++ imperium/src/commands/status.rs
@@ -31,7 +31,7 @@ pub enum WorkOrders {
 
     /// Get all details of a specific work order
     WorkOrder {
-        work_order_number: u32,
+        work_order_number: u64,
         level_of_detail: LevelOfDetail,
     },
 }
